Provide example of request JSON
I found that the request JSON has a minimum requirement.
I learned that I can not leave out the `jsonrpc`, `id` or `method` field. All are required as a minimum.
I get the following error if the `jsonrpc` or `method` fields are left out.
```
{
    "jsonrpc": "2.0",
    "id": 1,
    "error": {
        "code": -32600,
        "message": "Invalid Request"
    }
}
```
I get nothing back if I leave out the `id` field.